write/elf: move AttributesWriter to encoder module
